src/requirements.txt
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
idna==2.7
Naked==0.1.31
pycparser==2.19
pycrypto==2.6.1
pyOpenSSL==18.0.0
python-pkcs11==0.5.0
PyYAML==3.13
requests==2.20.0
shellescape==3.4.1
six==1.11.0
urllib3==1.24
RestrictedPython==4.0b4
python-dateutil==2.7.5
